Added 3 new modules
Added the Memory Module. It will ask for all the info, so you'll never get lost.
Added the Who's On First Module. It will simply display the list of words that the defuser needs. Might make it more pleasantly looking at some point.
Added the Complicated Wires. It will ask for the coloring of the wire, prioritizing red, blue or both. It will also ask for the LED and the star and will only ask for the batteries, parallel port and the serial number if necessary.