Added 2 new modules
Added the Morse Code Module. You can type whatever word in morse code and the app will then print it, then you can compare that word with any of the required words, to get the requires frequency.
Added the Wire Sequences Module
The app will ask for the color of the first wire and based on its ocurrences, it will tell if it needs to be cut or not.