Added 1 new module, fixed old ones
Added the Maze Module. The app will ask for any of the circles in the maze, and display the corresponding maze. Then, you can place the light and triangle on the maze, to make it easier to guide the defuser.
Fixed the Simon Says Module. Now you can place any number of colors and the app will then return how many and what colors you need to press.
Fixed the Complex Wires Module. It was telling to cut wires that didnt need cutting.