Added 2 new and last modules. Remade the entry
Added the Passwords Module. This module will ask for all the letters in all the positions, and show all the possible words everytime the expert inserts a new set of letters.
Added the Knob Module. This module will ask about the six most left LEDs of the module in the bomb, and based on that, this module will tell which position the knob needs to be.
Remade the startup. Now the App will start with a little bit of information about itself, and then, it will ask for important information of the bomb, like the serial number, how many batteries does it have and if it has a parallel port.