Remade and fix Memory Module.
Fixed the Memory Module, and simplified the code, so it will now work as intended and only ask for important info.
Added a new feature to the Complex Wires Module. Now it will show all the wire specifications you selected and keep them there in case there are multiple wires with that configuration