Wires Module Remade - Minor Code Simplification
The Wires Module was completely redone: now, instead of prompting questions for the defuser to answer, the expert will simply insert the cables in order top to bottom of the module in the bomb, and the app will automatically spell which wire to cut. It comes with a little wire module, so it can visually show the cables that the defuser is looking at.
In different modules, the code was simplified, so it wont consume as much.